Added file manager implementation for Windows

+++ Sierra/src/Core/Platform/Android/GameActivityScreen.h
@@ -29,7 +29,7 @@ namespace Sierra
 
         /* --- GETTER METHODS --- */
         [[nodiscard]] std::string_view GetName() const override { return name; };
-        [[nodiscard]] uint32 GetRefreshRate() const override { return refreshRate; };
+        [[nodiscard]] uint16 GetRefreshRate() const override { return refreshRate; };
         [[nodiscard]] ScreenOrientation GetOrientation() const override;
 
         [[nodiscard]] Vector2Int GetOrigin() const override { return origin; };
@@ -48,7 +48,7 @@ namespace Sierra
         Vector2UInt size = { 0, 0 };
         Vector2Int workAreaOrigin = { 0, 0 };
         Vector2UInt workAreaSize = { 0, 0 };
-        uint32 refreshRate = 0;
+        uint16 refreshRate = 0;
 
     };
