+++ Sierra/src/Core/Platform/Linux/X11Screen.h
@@ -32,7 +32,7 @@ namespace Sierra
 
         /* --- GETTER METHODS --- */
         [[nodiscard]] std::string_view GetName() const override { return name; };
-        [[nodiscard]] uint32 GetRefreshRate() const override { return refreshRate; };
+        [[nodiscard]] uint16 GetRefreshRate() const override { return refreshRate; };
         [[nodiscard]] ScreenOrientation GetOrientation() const override { return size.x >= size.y ? ScreenOrientation::Landscape : ScreenOrientation::Portrait; }
 
         [[nodiscard]] Vector2Int GetOrigin() const override { return origin; };
@@ -52,7 +52,7 @@ namespace Sierra
         Vector2UInt size = { 0, 0 };
         Vector2Int workAreaOrigin = { 0, 0 };
         Vector2UInt workAreaSize = { 0, 0 };
-        uint32 refreshRate = 0;
+        uint16 refreshRate = 0;
 
     };
